+++ src/serialization/gmetastorage_json.cpp
@@ -376,7 +376,7 @@ void G_API_CC GJsonStorageWriter::writeFundamental(const char * name, const GVar
 
 	JsonNodeType & newNode = this->addNode(name);
 
-	if(vtIsReal((uint16_t)vt)) {
+	if(vtIsReal(vt)) {
 		newNode = fromVariant<double>(v);
 	}
 	else {
@@ -384,7 +384,7 @@ void G_API_CC GJsonStorageWriter::writeFundamental(const char * name, const GVar
 			newNode = fromVariant<double>(v);
 		}
 		else {
-			if(vtIsUnsignedInteger((uint16_t)vt)) {
+			if(vtIsUnsignedInteger(vt)) {
 				newNode= fromVariant<unsigned long long>(v);
 			}
 			else {
